+++ canonicalwebteam/discourse/__init__.py
@@ -5,6 +5,8 @@ from canonicalwebteam.discourse.app import (  # noqa
     Category,  # noqa
     Events,  # noqa
 )
+from canonicalwebteam.discourse.cache import ResponseCache  # noqa
+from canonicalwebteam.discourse.exceptions import RateLimitedError  # noqa
 from canonicalwebteam.discourse.models import DiscourseAPI  # noqa
 from canonicalwebteam.discourse.parsers import (  # noqa
     DocParser,  # noqa